Add files via upload
These are both of the datasets created for demonstrated purposes only. 1 which is the dirty data in need to be dirty, by that i mean the dataset contains null values , columns are not identified correctly and rows contain duplicate values. while the other 2 is already dealt with data by removing duplicated and filling in null values. As I'm sure this is not the final dataset this should be sufficient to display the capabilities of our MVP.