FAB-121-ResearchCompetitors
A collation of findings on a key competitor, oracle analytics. The document provides summary merits and demerits of the application from a user perspective having explored the application thoroughly. In conjunction with Nick I hope to give a presentation later explaining the findings and how they may be relevant to our product