Add files via upload
A quick understanding of all of our available competitors, the next step is to acquire further insights onto what are these guys which are paid doing , and why are they all turning paid. this is a question we must anwser when specifying our scope, we must get our scope confirmed by brendan asap.