FAB-177/172/179-Ability-to-normalise-data-in-a-column/Outliers/Nulls
Changes made to fill_missing_values methods and remove_outliers methods to highlight changes in the uI requirements. normalise_data method added to allow data normalisation with two arguments. UI requirements document added to detail changes and requirements in the UI and give guide as to what is needed in UI.